app/requirements.txt
celery[py-amqp,redis]
python-dotenv